replace flake8 by ruff (#167)
* chore: Update dependencies and project configuration

* chore: Update .ruff.toml file

* chore: Update Ruff version to v0.5.6 and run linter and formatter

This commit updates the Ruff version to v0.5.6 in the .pre-commit-config.yaml file. It also adds the linter and formatter hooks for Ruff, with the "--fix" argument for the linter. This ensures that the code is properly formatted and follows the linting rules.

Note: This message does not include any issue references, tags, or author names.

* chore: Update Ruff version to v0.5.6, add linter and formatter hooks, and update dependencies

* chore: Update mermaid-py dependencies and add Ruff badge

This commit updates the dependencies for mermaid-py and adds a Ruff badge to the README.md file. The Ruff badge provides information about the latest version of Ruff being used in the project.

Note: This message does not include any issue references, tags, or author names.